--- a/common/decklist.cpp
+++ b/common/decklist.cpp
@@ -798,26 +798,50 @@ bool DeckList::deleteNode(AbstractDecklistNode *node, InnerDecklistNode *rootNod
 
 void DeckList::updateDeckHash()
 {
-    QStringList cardList;
-    QSet<QString> hashZones, optionalZones;
+    QMap<QByteArray, int> cardList;
+    QMap<QByteArray, int> sbCardList;
+    // all cards are sorted, yet as all cards are made lower case and SB: is added to all sideboard cards, they will always sort before mainboard cards
 
-    hashZones << DECK_ZONE_MAIN << DECK_ZONE_SIDE; // Zones in deck to be included in hashing process
-    optionalZones << DECK_ZONE_TOKENS;             // Optional zones in deck not included in hashing process
-
-    for (int i = 0; i < root->size(); i++) {
-        auto *node = dynamic_cast<InnerDecklistNode *>(root->at(i));
-        for (int j = 0; j < node->size(); j++) {
-            if (hashZones.contains(node->getName())) // Mainboard or Sideboard
-            {
-                auto *card = dynamic_cast<DecklistCardNode *>(node->at(j));
-                for (int k = 0; k < card->getNumber(); ++k) {
-                    cardList.append((node->getName() == DECK_ZONE_SIDE ? "SB:" : "") + card->getName().toLower());
-                }
+    for (AbstractDecklistNode *abstractNode : *root) {
+        auto *node = dynamic_cast<InnerDecklistNode *>(abstractNode);
+        QString zone = node->getName();
+        if (zone == DECK_ZONE_SIDE) {
+            for (AbstractDecklistNode *abstractNode : *node) {
+                auto *card = dynamic_cast<DecklistCardNode *>(abstractNode);
+                sbCardList[card->getName().toLower().toUtf8()] += card->getNumber();
+            }
+        } else if (zone == DECK_ZONE_MAIN) {
+            for (AbstractDecklistNode *abstractNode : *node) {
+                auto *card = dynamic_cast<DecklistCardNode *>(abstractNode);
+                cardList[card->getName().toLower().toUtf8()] += card->getNumber();
             }
         }
     }
-    cardList.sort();
-    QByteArray deckHashArray = QCryptographicHash::hash(cardList.join(";").toUtf8(), QCryptographicHash::Sha1);
+
+    QCryptographicHash hasher(QCryptographicHash::Sha1);
+    bool started = false;
+    for (auto i = sbCardList.cbegin(), end = sbCardList.cend(); i != end; ++i) {
+        for (int j = 0; j < i.value(); ++j) {
+            if (started) {
+                hasher.addData(";SB:");
+            } else {
+                hasher.addData("SB:");
+                started = true;
+            }
+            hasher.addData(i.key());
+        }
+    }
+    for (auto i = cardList.cbegin(), end = cardList.cend(); i != end; ++i) {
+        for (int j = 0; j < i.value(); ++j) {
+            if (started) {
+                hasher.addData(";");
+            } else {
+                started = true;
+            }
+            hasher.addData(i.key());
+        }
+    }
+    QByteArray deckHashArray = hasher.result();
     quint64 number = (((quint64)(unsigned char)deckHashArray[0]) << 32) +
                      (((quint64)(unsigned char)deckHashArray[1]) << 24) +
                      (((quint64)(unsigned char)deckHashArray[2] << 16)) +