--- a/tests/CMakeLists.txt
+++ b/tests/CMakeLists.txt
@@ -4,6 +4,7 @@ add_test(NAME expression_test COMMAND expression_test)
 
 add_test(NAME test_age_formatting COMMAND test_age_formatting)
 add_test(NAME password_hash_test COMMAND password_hash_test)
+add_test(NAME deck_hash_test COMMAND deck_hash_test)
 
 # Find GTest
 
@@ -11,6 +12,7 @@ add_executable(dummy_test dummy_test.cpp)
 add_executable(expression_test expression_test.cpp)
 add_executable(test_age_formatting test_age_formatting.cpp)
 add_executable(password_hash_test password_hash_test.cpp)
+add_executable(deck_hash_test ../common/decklist.cpp deck_hash_test.cpp)
 
 find_package(GTest)
 
@@ -41,6 +43,7 @@ if(NOT GTEST_FOUND)
   add_dependencies(expression_test gtest)
   add_dependencies(test_age_formatting gtest)
   add_dependencies(password_hash_test gtest)
+  add_dependencies(deck_hash_test gtest)
 endif()
 
 include_directories(${GTEST_INCLUDE_DIRS})
@@ -48,6 +51,7 @@ target_link_libraries(dummy_test Threads::Threads ${GTEST_BOTH_LIBRARIES})
 target_link_libraries(expression_test cockatrice_common Threads::Threads ${GTEST_BOTH_LIBRARIES} ${TEST_QT_MODULES})
 target_link_libraries(test_age_formatting Threads::Threads ${GTEST_BOTH_LIBRARIES} ${TEST_QT_MODULES})
 target_link_libraries(password_hash_test cockatrice_common Threads::Threads ${GTEST_BOTH_LIBRARIES} ${TEST_QT_MODULES})
+target_link_libraries(deck_hash_test cockatrice_common Threads::Threads ${GTEST_BOTH_LIBRARIES} ${TEST_QT_MODULES})
 
 add_subdirectory(carddatabase)
 add_subdirectory(loading_from_clipboard)