## Related Ticket(s)
- Related #5025

## Short roundup of the initial problem
There are no tests that verify the deck hasher remains the same.

## What will change with this Pull Request?
- adds tests
- tries to improve performance