@ebbit1q 2024-04-29
the tests are adequate but I am not satisfied with my attempt at improving performance

before:
```
[==========] Running 3 tests from 1 test suite.
[----------] Global test environment set-up.
[----------] 3 tests from DeckHashTest
[ RUN      ] DeckHashTest.RepeatTest
[       OK ] DeckHashTest.RepeatTest (8168 ms)
[ RUN      ] DeckHashTest.NumberTest
[       OK ] DeckHashTest.NumberTest (616 ms)
[ RUN      ] DeckHashTest.UniquesTest
[       OK ] DeckHashTest.UniquesTest (2831 ms)
[----------] 3 tests from DeckHashTest (11617 ms total)

[----------] Global test environment tear-down
[==========] 3 tests from 1 test suite ran. (11617 ms total)
[  PASSED  ] 3 tests.
```

after:
```
[==========] Running 3 tests from 1 test suite.
[----------] Global test environment set-up.
[----------] 3 tests from DeckHashTest
[ RUN      ] DeckHashTest.RepeatTest
[       OK ] DeckHashTest.RepeatTest (4033 ms)
[ RUN      ] DeckHashTest.NumberTest
[       OK ] DeckHashTest.NumberTest (93 ms)
[ RUN      ] DeckHashTest.UniquesTest
[       OK ] DeckHashTest.UniquesTest (6718 ms)
[----------] 3 tests from DeckHashTest (10845 ms total)

[----------] Global test environment tear-down
[==========] 3 tests from 1 test suite ran. (10845 ms total)
[  PASSED  ] 3 tests.
```

we see the total improvement over all 3 tests is negligible because while the first two are faster the last is much slower.