@ebbit1q 2024-04-29
I agree, I'm thinking of whether to make it a per node limit (easy to implement), or a total card limit, but I think it should be a limit on the deck itself, not just the hash.