@ebbit1q 2025-10-03
I cast advanced necromancy: revive useful part of ancient pull request